refactor: extract universal v-model codegen code and update weex v-model codegen
