+++ test/weex/compiler/v-model.spec.js
@@ -14,28 +14,16 @@ describe('compile v-model', () => {
   it('should compile other component with whole $event as the value', () => {
     const { render, staticRenderFns, errors } = compile(`<div><foo v-model="x" /></div>`)
     expect(render).not.toBeUndefined()
-    expect(render).toMatch(strToRegExp(`attrs:{"value":(x)}`))
-    expect(render).toMatch(strToRegExp(`on:{"input":function($event){x=$event}}`))
-    expect(staticRenderFns).toEqual([])
-    expect(errors).toEqual([])
-  })
-
-  it('should compile with lazy modifier', () => {
-    const { render, staticRenderFns, errors } = compile(`<div><foo v-model.lazy="x" /></div>`)
-    expect(render).not.toBeUndefined()
-    expect(render).toMatch(strToRegExp(`attrs:{"value":(x)}`))
-    expect(render).toMatch(strToRegExp(`on:{"change":function($event){x=$event}}`))
+    expect(render).toMatch(strToRegExp(`model:{value:(x),callback:function (value) {x=value}}`))
     expect(staticRenderFns).toEqual([])
     expect(errors).toEqual([])
   })
 
   it('should compile with trim modifier for modelable native component', () => {
-    const { render, staticRenderFns, errors } = compile(`<div><input v-model.trim="x" /><foo v-model.trim="y" /></div>`)
+    const { render, staticRenderFns, errors } = compile(`<div><input v-model.trim="x" /></div>`)
     expect(render).not.toBeUndefined()
     expect(render).toMatch(strToRegExp(`attrs:{"value":(x)}`))
-    expect(render).toMatch(strToRegExp(`attrs:{"value":(y)}`))
     expect(render).toMatch(strToRegExp(`on:{"input":function($event){x=$event.target.attr.value.trim()}}`))
-    expect(render).toMatch(strToRegExp(`on:{"input":function($event){y=$event}}`))
     expect(staticRenderFns).toEqual([])
     expect(errors).toEqual([])
   })
